# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/providers": {
      "post": {
        "operationId": "ProvidersController_create",
        "summary": "Add a new AI provider configuration (OpenAI / Claude / Gemini)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created; apiKeyPreview shows a masked key"
          },
          "400": {
            "description": "Validation failed (missing/short apiKey, invalid enum, etc.)"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "get": {
        "operationId": "ProvidersController_findAll",
        "summary": "List the current user's configured AI providers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of provider configs (keys never exposed)"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/providers/{id}": {
      "get": {
        "operationId": "ProvidersController_findOne",
        "summary": "Get a single provider configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider found"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "ProvidersController_update",
        "summary": "Edit a provider configuration (label, model, key rotation, etc.)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "ProvidersController_remove",
        "summary": "Delete a provider configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider removed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/providers/{id}/enable": {
      "patch": {
        "operationId": "ProvidersController_enable",
        "summary": "Enable a provider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider enabled"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/providers/{id}/disable": {
      "patch": {
        "operationId": "ProvidersController_disable",
        "summary": "Disable a provider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider disabled"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/providers/default": {
      "get": {
        "operationId": "ProvidersController_getDefault",
        "summary": "Get the default AI provider configuration",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Default provider found"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "No default provider configured"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/providers/{id}/default": {
      "patch": {
        "operationId": "ProvidersController_setDefaultPatch",
        "summary": "Set provider as default",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider set as default"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Provider not found or not owned by this user"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "ProvidersController_setDefaultPost",
        "summary": "Set provider as default (POST alias)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider set as default"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "AI Providers"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/search/query": {
      "post": {
        "operationId": "SearchController_search",
        "summary": "Run an AI-assisted web search query",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Search results (served from cache when available within the cache window)"
          },
          "400": {
            "description": "Validation failed on the request body"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Daily usage limit reached for the current plan"
          }
        },
        "tags": [
          "Web Search"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/search": {
      "post": {
        "operationId": "SearchController_searchRoot",
        "summary": "Run an AI-assisted web search query (root alias)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchQueryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Search results (served from cache when available within the cache window)"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Web Search"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/search/history": {
      "get": {
        "operationId": "SearchController_history",
        "summary": "Paginated history of the user's past searches",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of past searches, most recent first"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Web Search"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/search/recent": {
      "get": {
        "operationId": "SearchController_recent",
        "summary": "Most recent distinct searches",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of recent distinct search queries"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Web Search"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/search/suggestions": {
      "get": {
        "operationId": "SearchController_suggestions",
        "summary": "Autocomplete suggestions from the user's own search history",
        "parameters": [
          {
            "name": "prefix",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of suggested query strings"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "tags": [
          "Web Search"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/admin/providers/{id}/disable": {
      "patch": {
        "operationId": "AdminController_disableProvider",
        "summary": "Force-disable any user's provider (e.g. reported abuse, leaked key)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider disabled"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "Provider not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/providers/{id}/enable": {
      "patch": {
        "operationId": "AdminController_enableProvider",
        "summary": "Re-enable any user's provider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider enabled"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "Provider not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/providers/{id}": {
      "delete": {
        "operationId": "AdminController_deleteProvider",
        "summary": "Force-delete any user's provider configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider removed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Authenticated but not an ADMIN"
          },
          "404": {
            "description": "Provider not found"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT Backend API",
    "description": "Production-ready backend for the EchoGPT Chrome Extension: auth, subscriptions, AI provider management, chat, web search, and admin analytics.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "OPENAI",
              "CLAUDE",
              "GEMINI"
            ]
          },
          "label": {
            "type": "string",
            "description": "Friendly label, e.g. \"Work OpenAI key\""
          },
          "apiKey": {
            "type": "string",
            "description": "Plaintext API key - encrypted at rest, never returned again"
          },
          "model": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string",
            "description": "Override base URL (self-hosted / proxy endpoints)"
          },
          "isDefault": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "name",
          "apiKey"
        ]
      },
      "UpdateProviderDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "apiKey": {
            "type": "string",
            "description": "Provide only to rotate the stored key"
          },
          "model": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
      "SearchQueryDto": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "example": "latest NestJS release notes"
          }
        },
        "required": [
          "query"
        ]
      },
...
    }
  }
}